requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiogram==3.2.0
python-dotenv==1.0.0
pydantic==2.5.0
asyncpg==0.29.0
SQLAlchemy==2.0.23
alembic==1.12.1
httpx==0.25.2
python-jose[cryptography]==3.3.0
pytest==7.4.3
python-multipart==0.0.6
pydantic-settings==2.1.0